requirements.txt
fastapi
uvicorn
openai
pandas
gradio
pyyaml
